+++ CHANGELOG.md
@@ -13,4 +13,12 @@ We do not give estimated times for completion on `Accepted` Proposals.
 
 ### v1.0.0 - 2015-01-20
 
-- Initial release.
+- Can create, update, delete files.
+- Can email files.
+- Can set private/public.
+- Can share files.
+- Can download files.
+- Has blade call `@media('id', 'download|thumbnail')`
+- Has blade call `@thumbnail('id', [ 'options' ], 'default')`
+- Can add tags.
+- Can create config styles.